+++ lib/src/constants.dart
@@ -39,6 +39,7 @@ const maxInlineSize = 32768;
 /// Default DM relays used when user has no relay list configured
 const recommendedDmRelays = [
   'wss://auth.nostr1.com',
+  'wss://relay.nmail.li',
   'wss://nostr-01.uid.ovh',
   'wss://nostr-02.uid.ovh',
 ];
@@ -46,6 +47,7 @@ const recommendedDmRelays = [
 /// Default Blossom servers used when user has no server list configured
 const recommendedBlossomServers = [
   'https://blossom.yakihonne.com',
+  'https://blossom.nmail.li',
   'https://blossom-01.uid.ovh',
   'https://blossom-02.uid.ovh',
   'https://blossom.primal.net',