+++ test/integration/client/account_isolation_test.dart
@@ -186,7 +186,11 @@ void main() {
         privkey: aliceAccount.privateKey!,
       );
 
-      client = NostrMailClient(ndk: ndk, db: db, defaultDmRelays: [relay.url]);
+      client = await NostrMailClient.create(
+        ndk: ndk,
+        db: db,
+        defaultDmRelays: [relay.url],
+      );
     });
 
     void addBob() {